refactor: move FileModule to commons, remove unused functions
